Fix bug preventing page scroll. Breaks marker pos.
This commit fixes a bug where you could only scroll the code view
element, not the whole page. This meant that you couldn't scroll the
page down past the GitHub header (if your page loaded in the default,
scrolled-to-the-beggining-of-the-page mode.)

HOWEVER, this also breaks marker positions. Because we no longer bound
the size of the code view element (and scroll the overflow,) the code
marker container matches the size of the code edit window, and so
marker layout is idempotent with code position; that is, the markers
are places on the page next to the line of code their marking. This
does us no good, since to see the shortcut to that line of code, we
need to... manually scroll to that line of code.

Working on fixing that.