Display line marks in container fixed to viewport
This fixes the bug introduced in #09ece07df0 where line marks were
distributed over the length of the code blob container (and thus you
had to scroll to the line being marked to see the line marking.)

Now, the line markings container is sized to the browser window so that
it's possible to see all the marking in one window.

Further, when the code blob container fills the screen, the mark
container is fixed in place to the viewport (so it's always on screen
in the same place.) Otherwise (when, for example, we haven't scrolled
down so the code blob container fills the screen,) the code marks
container is at the top of the code blob container.